Updated the way the style conditions work for the different device sizes. Url details are being updated uncessarily So moved the function used in renderengine to a util
