+++ ui-app/client/src/App/App.tsx
@@ -77,47 +77,47 @@ function setDeviceType() {
 
 	if (size >= (StyleResolutionDefinition.get(StyleResolution.WIDE_SCREEN)?.minWidth ?? 1281)) {
 		newDevices[StyleResolution.WIDE_SCREEN] = true;
+		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
+		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
 	} else if (
 		size >= (StyleResolutionDefinition.get(StyleResolution.DESKTOP_SCREEN)?.minWidth ?? 1025)
 	) {
 		newDevices[StyleResolution.DESKTOP_SCREEN_ONLY] = true;
 		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
-		newDevices[StyleResolution.WIDE_SCREEN] = true;
+		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
 	} else if (
 		size >=
 		(StyleResolutionDefinition.get(StyleResolution.TABLET_LANDSCAPE_SCREEN)?.minWidth ?? 961)
 	) {
 		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN_ONLY] = true;
 		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
-		newDevices[StyleResolution.WIDE_SCREEN] = true;
+		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
 	} else if (
 		size >=
 		(StyleResolutionDefinition.get(StyleResolution.TABLET_POTRAIT_SCREEN)?.minWidth ?? 641)
 	) {
 		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN_ONLY] = true;
 		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
-		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
-		newDevices[StyleResolution.WIDE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
 	} else if (
 		size >=
 		(StyleResolutionDefinition.get(StyleResolution.MOBILE_LANDSCAPE_SCREEN)?.minWidth ?? 481)
 	) {
 		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN_ONLY] = true;
 		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
-		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
-		newDevices[StyleResolution.WIDE_SCREEN] = true;
+		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
 	} else {
 		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN_ONLY] = true;
 		newDevices[StyleResolution.MOBILE_POTRAIT_SCREEN] = true;
-		newDevices[StyleResolution.MOBILE_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.TABLET_POTRAIT_SCREEN] = true;
-		newDevices[StyleResolution.TABLET_LANDSCAPE_SCREEN] = true;
-		newDevices[StyleResolution.DESKTOP_SCREEN] = true;
-		newDevices[StyleResolution.WIDE_SCREEN] = true;
 	}
 
 	let devicesString = JSON.stringify(newDevices);
