+++ ui-app/client/src/Engine/RenderEngineContainer.tsx
@@ -11,6 +11,7 @@ import {
 import * as getPageDefinition from './../definitions/getPageDefinition.json';
 import { runEvent } from '../components/util/runEvent';
 import Page from '../components/Page';
+import { processLocation } from '../util/locationProcessor';
 
 export const RenderEngineContainer = () => {
 	const location = useLocation();
@@ -19,7 +20,11 @@ export const RenderEngineContainer = () => {
 	const [pageDefinition, setPageDefinition] = useState<any>();
 
 	useEffect(() => {
-		let { pageName } = processLocation(location);
+		const details = processLocation(location);
+		setData(`${STORE_PREFIX}.urlDetails`, details);
+
+		let { pageName } = details;
+
 		if (!pageName)
 			pageName = getDataFromPath(`${STORE_PREFIX}.application.properties.defaultPage`, []);
 		let pDef = getDataFromPath(`${STORE_PREFIX}.pageDefinition.${pageName}`, []);
@@ -93,42 +98,3 @@ export const RenderEngineContainer = () => {
 		);
 	}
 };
-
-export function processLocation(location: Location) {
-	const details: any = { queryParameters: {} };
-
-	if (location.search) {
-		details.queryParameters = location.search
-			.split('&')
-			.map(e => {
-				if (e.startsWith('?')) e = e.substring(1);
-				const two = e.split('=');
-				if (two.length === 0) return undefined;
-				if (two.length === 1) return { key: decodeURIComponent(two[0]), value: '' };
-				return { key: decodeURIComponent(two[0]), value: decodeURIComponent(two[1]) };
-			})
-			.reduce((a: any, c) => {
-				if (!c) return a;
-
-				a[c.key] = c.value;
-				return a;
-			}, {});
-	}
-
-	if (location.pathname) {
-		const pathParts = location.pathname.split('/').filter(e => e !== '');
-
-		const ind = pathParts.indexOf('page');
-		if (ind === -1) {
-			details.pathParts = pathParts;
-			details.pageName = pathParts[0];
-		} else {
-			if (ind > 0) details.appName = pathParts[0];
-			if (ind > 1) details.clientCode = pathParts[1];
-			details.pageName = pathParts[ind + 1];
-			details.pathParts = pathParts.slice(ind + 1);
-		}
-	}
-	setData(`${STORE_PREFIX}.urlDetails`, details);
-	return details;
-}
